+++ clients/android/NewsBlur/res/values/theme.xml
@@ -163,33 +163,4 @@
         <item name="fontFamily">@font/whitney</item>
         <item name="circleProgressIndicator">@style/circleProgressIndicator</item>
     </style>
-
-    <style name="NewsBlurTheme.Translucent" parent="NewsBlurTheme">
-        <item name="android:windowBackground">@android:color/transparent</item>
-        <item name="android:colorBackgroundCacheHint">@null</item>
-        <item name="android:windowIsTranslucent">true</item>
-        <item name="android:windowAnimationStyle">@android:style/Animation</item>
-    </style>
-
-    <style name="NewsBlurBlackTheme.Translucent" parent="NewsBlurBlackTheme">
-        <item name="android:windowBackground">@android:color/transparent</item>
-        <item name="android:colorBackgroundCacheHint">@null</item>
-        <item name="android:windowIsTranslucent">true</item>
-        <item name="android:windowAnimationStyle">@android:style/Animation</item>
-    </style>
-
-    <style name="NewsBlurDarkTheme.Translucent" parent="NewsBlurDarkTheme">
-        <item name="android:windowBackground">@android:color/transparent</item>
-        <item name="android:colorBackgroundCacheHint">@null</item>
-        <item name="android:windowIsTranslucent">true</item>
-        <item name="android:windowAnimationStyle">@android:style/Animation</item>
-    </style>
-
-    <style name="Theme.Translucent" parent="Theme.MaterialComponents.NoActionBar">
-        <item name="android:background">@android:color/transparent</item>
-        <item name="android:windowBackground">@android:color/transparent</item>
-        <item name="android:colorBackgroundCacheHint">@null</item>
-        <item name="android:windowIsTranslucent">true</item>
-        <item name="android:windowAnimationStyle">@android:style/Animation</item>
-    </style>
 </resources>
