--- clients/android/NewsBlur/src/com/newsblur/network/domain/ShareExternalStoryResponse.kt
@@ -1,6 +1,0 @@
-package com.newsblur.network.domain
-
-class ShareExternalStoryResponse : NewsBlurResponse() {
-
-    override fun isError() = code != 0
-}
